Broke visualization into separate components
