+++ visualization/app/.eslintrc.js
@@ -1,39 +1,16 @@
 module.exports = {
-  env: {
-    browser: true,
-    es6: true
-  },
-  extends: ['eslint:recommended', 'plugin:react/recommended'],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly'
-  },
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true
-    },
-    ecmaVersion: 2018,
-    sourceType: 'module'
-  },
-  plugins: ['react'],
+  root: true,
+  extends: "universe/web",
   rules: {
-    indent: ['error', 2],
-    'linebreak-style': ['error', 'unix'],
-    quotes: ['error', 'single'],
-    semi: ['error', 'always']
+    "no-console": process.env.NODE_ENV === "production" ? "error" : "off",
+    "no-debugger": process.env.NODE_ENV === "production" ? "error" : "off",
+  },
+  globals: {
+    process: true,
   },
   settings: {
     react: {
-      createClass: 'createReactClass',
-      pragma: 'React',
-      version: 'detect',
-      flowVersion: '0.53'
+      version: "detect",
     },
-    propWrapperFunctions: [
-      'forbidExtraProps',
-      { property: 'freeze', object: 'Object' },
-      { property: 'myFavoriteWrapper' }
-    ],
-    linkComponents: ['Hyperlink', { name: 'Link', linkAttribute: 'to' }]
-  }
+  },
 };
