+++ visualization/app/package.json
@@ -14,7 +14,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "format": "prettier-eslint --write \"+(src|test)/**/*.{js,json}\""
+    "format": "prettier-eslint --write \"+(src|test)/**/*.{js,json}\"",
+    "lint": "eslint ."
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -32,6 +33,7 @@
     ]
   },
   "devDependencies": {
+    "eslint-config-universe": "^1.0.7",
     "eslint-plugin-react": "^7.14.3",
     "prettier": "^1.18.2",
     "prettier-eslint-cli": "^5.0.0"
