+++ visualization/app/src/components/Chills.js
@@ -1,0 +1,7 @@
+import React, { Component } from 'react';
+
+export default class Chills extends Component {
+  render() {
+    return <div>Chills</div>;
+  }
+}
