+++ visualization/app/src/components/Colors/index.js
@@ -1,0 +1,7 @@
+import React, { Component } from 'react';
+
+export default class Colors extends Component {
+  render() {
+    return <div>Colors</div>;
+  }
+}
