+++ visualization/app/src/components/Synesthesia/Bee.js
@@ -1,20 +1,13 @@
 import React from 'react';
 
-const Bee = ({
-  className = '',
-  fill = 'white',
-  stroke = 'black',
-  style = {},
-  width = '100%'
-}) => (
+const Bee = ({ className = '', fill = 'white', stroke = 'black', style = {}, width = '100%' }) => (
   <svg
     className={className}
     height={width}
     style={style}
     viewBox="0 0 600 600"
     width={width}
-    xmlns="http://www.w3.org/2000/svg"
-  >
+    xmlns="http://www.w3.org/2000/svg">
     <path
       fill={stroke}
       d="m 260.54018,75.61161 c -5.42306,0.0481 -10.84715,0.6481 -16.09375,1.6875 -6.09251,3.33914 -16.15661,4.79344 -18.65625,14.28125 -8.95848,0.6667 -18.041,3.50165 -27.21875,3.46875 -9.29648,0.0642 -18.20081,2.24739 -26.59375,6 -8.40125,6.07554 -16.62141,12.37678 -25.34375,18.0625 -9.0873,3.47942 -7.39324,16.80155 -17.8125,14.9375 -8.45167,2.45166 -16.66953,7.63957 -21.90625,14.78125 -3.07061,9.06527 -10.174468,12.09001 -17.562501,17.25 -6.08585,6.80923 -17.420306,13.20198 -11.3125,24.3125 -2.133027,8.88588 -13.684699,13.28111 -14.21875,23.71875 4.76817,10.48159 -10.934365,8.99088 -9.46875,19.5 1.592176,8.88606 -1.553412,16.0755 -5.8125,23.53125 1.747728,7.98726 -0.01121,14.27063 -6.125,19.75 -4.640791,8.6402 5.005981,19.10488 -0.34375,26.78125 -2.634426,9.55892 -2.157609,19.96056 -0.6875,29.71875 3.694737,9.07811 5.766819,18.65109 9.28125,27.8125 3.967346,8.98917 8.393904,17.70335 18.59375,20.6875 2.719206,8.49925 13.422217,11.39914 21.46875,8.28125 10.570201,-4.41606 1.915245,8.8987 3.28125,14.09375 0.862599,8.52587 4.674688,17.58746 7.781251,23.9375 5.33891,5.76197 11.12194,11.88522 17.28125,16.96875 6.85549,6.38328 17.16231,7.33997 22.21875,15.625 5.07948,7.66924 17.71924,2.93317 21.1875,10.375 5.9912,8.57363 16.80035,3.34642 24.5625,7.875 8.46966,5.50031 17.99481,-1.32724 26.1875,2.59375 7.62679,4.14771 17.24712,3.73983 25.40625,1.53125 6.12591,-3.9614 8.78084,-0.34641 7.46875,6.59375 0.84387,8.45335 -1.48463,16.96495 -0.21875,25.1875 -1.26655,9.28686 4.09962,17.42644 13.90625,17.78125 7.41336,0.24444 16.70005,-1.42962 19.9375,-9.09375 1.81541,-10.27233 1.93106,-20.93871 4.21875,-31.21875 2.35633,-9.18153 3.21634,-18.90259 10.0625,-26.21875 8.14262,-11.02647 18.82386,-23.36049 16.40625,-38.09375 4.67553,-4.64486 14.66634,-8.9371 17.6875,-17.3125 7.58112,-9.40411 11.40965,-21.18129 18.53125,-30.84375 8.41102,3.01115 17.51385,6.85889 26.875,5.875 8.73405,-1.37018 17.48011,-4.12061 25.28125,-8.15625 3.6147,-10.30817 14.24773,-5.12515 22.3125,-9.8125 10.55472,-2.44906 16.22734,-11.72686 22.21875,-19.875 4.58408,-7.34509 11.92315,-14.30128 12.53125,-23.75 2.32873,-9.87896 2.82428,-20.77243 0.40625,-30.65625 -4.47886,-7.69495 2.48416,-5.73759 8.03125,-5.625 9.50853,-0.0975 19.3435,-1.40457 28.875,-1.96875 6.9735,-4.19836 17.05836,0.65174 23.21875,-7.21875 7.16108,-2.82014 15.92036,-3.49995 21.375,-10.46875 11.35794,-2.71286 16.6322,-14.44033 20.46875,-24.375 3.87919,-9.2437 5.98223,-19.29853 6.59375,-29.1875 -1.78717,-8.7762 -0.81926,-19.59055 -8.1875,-26.09375 -0.51817,-12.72059 -6.85042,-25.37068 -17.21875,-32.96875 -2.56322,-9.38911 -9.42608,-17.69028 -18.75,-21.375 -2.5186,-8.06486 -17.05469,-2.43968 -15.3125,-10.53125 -4.25638,-7.45909 -11.76556,-14.02892 -19.125,-17.6875 -7.36478,-3.70073 -16.87499,-0.008 -23.1875,-4.5 -7.21567,-6.47844 -17.28013,-7.57137 -26.375,-9.40625 -8.31882,-0.44585 -16.70135,-3.97553 -24.40625,-4.25 -5.96227,-3.66818 -15.72359,-9.02182 -21.5,-1.8125 -6.10103,-1.41179 -14.15787,-9.78007 -23.1875,-7.09375 -9.47654,1.2381 -19.05316,-3.59755 -28.84375,-2.34375 -7.5639,-3.02353 -16.3723,6.27315 -22.625,3.40625 -9.17707,-2.86672 -20.13893,-3.29146 -29.21875,0.1875 -8.24346,-3.31225 -17.30531,-4.73649 -26.34375,-4.65625 z"
