+++ visualization/app/src/components/Synesthesia/Synesthesia.css
