+++ visualization/app/src/components/index.js
@@ -1,0 +1,5 @@
+import Chills from './Chills';
+import Colors from './Colors';
+import Synesthesia from './Synesthesia';
+
+export { Chills, Colors, Synesthesia };
