+++ visualization/app/src/constants/ChoiceTypes.js
@@ -5,5 +5,5 @@ export const CHOICE_EMOTION_HAPPINESS = 'CHOICE_EMOTION_HAPPINESS';
 export default {
   CHOICE_COLOR,
   CHOICE_EMOTION_ANGER,
-  CHOICE_EMOTION_HAPPINESS
+  CHOICE_EMOTION_HAPPINESS,
 };
